membuat crud masterdata

+++ resources/views/document/index.blade.php
@@ -28,7 +28,14 @@
 
               <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal" data-whatever="@getbootstrap">Tambah</button>
 
-              <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+              @if (session('status'))
+                  <div class="alert alert-success">
+                      {{ session('status') }}
+                  </div>
+              @endif
+
+<!-- Content tambah modal -->
+            <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
               <div class="modal-dialog" role="document">
         <div class="modal-content">
           <div class="modal-header">
@@ -38,17 +45,25 @@
             </button>
           </div>
           <div class="modal-body">
-            <form method="POST" action="{{action('DocumentController@store')}}" >
+            <form method="POST" action="{{action('DocumentController@store')}}" id="editform" >
               {{csrf_field()}}
 
               <div class="form-group">
                 <label for="id" class="col-form-label">Id:</label>
-                <input name="id" type="number" class="form-control" id="id"placeholder="Masukkan Id">
+                <input name="id" type="number" class="form-control @error('id')is-invalid @enderror" id="id"placeholder="Masukkan Id" value="{{old('id')}}">
+                @error('id')
+                  <div class="invalid-feedback">{{$message}}</div>
+                @enderror
               </div>
+
               <div class="form-group">
                 <label for="document" class="col-form-label">Nama Document:</label>
-                <input name="document" type="text" class="form-control" id="document" placeholder="Masukkan document">
+                <input name="document" type="text" class="form-control @error('id')is-invalid @enderror" id="document" placeholder="Masukkan document" value="{{old('id')}}">
+                @error('document')
+                  <div class="invalid-feedback">{{$message}}</div>
+                @enderror
               </div>
+
               <div class="modal-footer">
                 <button type="submit" class="btn btn-primary">Buat</button>
                 <button type="button" class="btn btn-danger" data-dismiss="modal">Batal</button>
@@ -59,7 +74,51 @@
         </div>
       </div>
     </div>
+<!-- End Content Tambah modal -->
 
+<!-- Content Edit modal -->
+    @foreach($document as $dcm)
+      <div class="modal fade" id="editModal-{{$dcm->id}}" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+      <div class="modal-content">
+          <div class="modal-header">
+                      <h2 class="modal-title" id="exampleModalLabel">Tambah Document</h2>
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                              <span aria-hidden="true">&times;</span>
+            </button>
+            </div>
+              <div class="modal-body">
+              <form method="POST" action="{{action('DocumentController@edit')}}" id="editform" >
+                {{csrf_field()}}
+
+              <div class="form-group">
+                <label for="id" class="col-form-label">Id:</label>
+                  <input name="id" type="number" class="form-control @error('id')is-invalid @enderror" id="id"placeholder="Masukkan Id" value="{{$dcm->id}}">
+                        @error('id')
+                          <div class="invalid-feedback">{{$message}}</div>
+                        @enderror
+              </div>
+
+                <div class="form-group">
+                  <label for="document" class="col-form-label">Nama Document:</label>
+                  <input name="document" type="text" class="form-control @error('id')is-invalid @enderror" id="document" placeholder="Masukkan document" value="{{$dcm->document}}">
+                          @error('document')
+                            <div class="invalid-feedback">{{$message}}</div>
+                          @enderror
+                </div>
+
+                <div class="modal-footer">
+                          <button type="submit" class="btn btn-primary">Ubah</button>
+                          <button type="button" class="btn btn-danger" data-dismiss="modal">Batal</button>
+                </div>
+
+            </form>
+          </div>
+        </div>
+      </div>
+    </div>
+@endforeach
+<!-- End Content Edit modal -->
 
           <table id="example" class="display" style="width:100%">
 
@@ -78,8 +137,8 @@
               <td>{{$dcm->id}}</td>
               <td >{{$dcm->document}}</td>
                 <td class="text-center">
-                    <a href="" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal" data-whatever="@getbootstrap">edit</a>
-                    <a href="document.index.destroy{{$dcm->id }}" class="btn btn-danger">delete</a>
+                    <a href="" class="btn btn-primary" data-toggle="modal" data-target="#editModal-{{$dcm->id}}" data-whatever="@getbootstrap">edit</a>
+                    <a href="document.index.destroy{{$dcm->id }}" class="btn btn-danger" onclick="return confirm('Apakah anda yakin ingin mengapus data ?')" >delete</a>
                 </td>
             </tr>
           @endforeach
